[refactor]:Main #89 프로필 수정하기 Dto 생성
