+++ MainService/src/main/java/com/kusitms/mainservice/domain/mypage/dto/response/MyPageResponseDto.java
@@ -1,4 +1,4 @@
-package com.kusitms.mainservice.domain.mypage.dto;
+package com.kusitms.mainservice.domain.mypage.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
