+++ MainService/src/main/java/com/kusitms/mainservice/domain/mypage/dto/response/MyPageUserResponseDto.java
@@ -1,4 +1,4 @@
-package com.kusitms.mainservice.domain.mypage.dto;
+package com.kusitms.mainservice.domain.mypage.dto.response;
 
 import com.kusitms.mainservice.domain.user.domain.User;
 import lombok.Builder;
